requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
django-ckeditor>=6.7.0
django-taggit>=5.0.0

